proyecto_regionalGA4/requirements.txt
cmake>=3.24.1.1
cachetools>=4.2.2
certifi>=2020.12.5
cffi>=1.14.5 
chardet>=4.0.0
click>=7.1.2
cryptography>=3.4.7
example>=0.1.0 #new
MarkupSafe==2.1.1
protobuf==3.20.1 #new
flask==2.0.1
Flask-Cors>=3.0.10
google-api-core>=1.26.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0
Jinja2==3.0
oauth2client==4.1.3
packaging==20.9
protobuf>=3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.2.2
google-analytics-data==0.12.1
gunicorn>=20.0.1
pandas>=1.3.2
google~=3.0.0
mysql-connector-python~=8.0.27
google-auth==2.3.3
DateTime~=4.3
lxml~=4.8.0
gunicorn~=20.0.1
html5lib~=1.1
python-dotenv~=0.19.2
sendgrid~=6.4.6
google-api-python-client~=2.63.0
SQLAlchemy~=1.4.41
